tidy(app): move sqlite-specific objects to sqlite file
